requirements.txt
langchain-google-genai
langchain-groq
python-dotenv
langchain-chroma
langchain-voyageai
langchain-text-splitters
langchain-community
langgraph
langgraph-checkpoint-postgres
chromadb
rank-bm25
cohere
supabase
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
tavily-python
colorama